requirements.txt
selenium~=4.34.1
pytest~=8.4.1
colorama~=0.4.6
requests~=2.32.4
PyYAML~=6.0.2
dotenv~=0.9.9
python-dotenv~=1.1.1
webdriver-manager~=4.0.2